pyproject.toml
[tool.poetry]
name = "mydream-social-backend"
version = "0.1.0"
description = "Backend for social network"
authors = ["dreamisisteam"]
readme = "README.md"
package-mode=false


[tool.poetry.dependencies]
python = "^3.12"
numpy = "^2.1.2"
fastapi = {extras = ["standard"], version = "^0.115.2"}
pydantic = "^2.9.2"
pydantic-settings = "^2.6.0"
taskiq = "^0.11.7"
taskiq-redis = "^1.0.2"
poetry = "^1.8.4"
tortoise-orm = {extras = ["asyncpg"], version = "^0.21.7"}
aerich = "^0.7.2"
pyjwt = "^2.9.0"
redis = "^5.1.1"
taskiq-fastapi = "^0.3.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.7.0"
pre-commit = "^4.0.1"


[tool.aerich]
tortoise_orm = "database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./src"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
